New pressure gauge, fix errors cause by air

+++ src/main/java/me/desht/pneumaticcraft/client/render/tileentity/RenderAirCannon.java
@@ -20,7 +20,7 @@ public class RenderAirCannon extends AbstractModelRenderer<TileEntityAirCannon> 
 
     @Override
     void renderModel(TileEntityAirCannon te, float partialTicks) {
-        float angle = (float) RenderUtils.rotateMatrixByMetadata(te.getBlockMetadata());
+        float angle = RenderUtils.rotateMatrixByMetadata(te.getBlockMetadata());
         float rotationAngle = te.rotationAngle - angle + 180F;
         this.model.renderModel(0.0625F, rotationAngle, te.heightAngle);
     }
